add docs for scala
